Update QueryFinishPending test for SharedInputScan
The intent of these tets are to test fault injection for plans with
SharedInputScan over a Sort. Prior to the postgres window merge, Planner
generated SharedInputScan for the window functions; now it does not.

To fulfill the intent; change the query to use WITH clause and ORDER BY
such that Planner generated SharedInputScan over a Sort node.

Signed-off-by: Ekta Khanna <ekhanna@pivotal.io>