ORCA now mimics planner when it comes to empty stats
When there is no stats available for any table, ORCA was treating it as an
empty table while planning. On the other hand planner is utilizing a guc
`gp_enable_relsize_collection` to obtain the estimated size of the table, but
no other statistics. This commit enables ORCA to have the same behavior as
planner.

Signed-off-by: Sambitesh Dash <sdash@pivotal.io>