Fix analyze table with more than 800 numeric columns
While analyzing a table, non-text varlen cols would be replaced with null if
detoasted tuple size is more than WIDTH_THRESHOLD. We also create a separate
indicator column to mark such tuple.  We don't need to do this for numeric
columns. The maximum size of numeric cols is less than WIDTH_THRESHOLD.

With this change, we can now analyze a table with more than 830 numeric cols.